requirements.txt
PyYAML==6.0.2
pydantic==2.8.2